:recycle: Move ParagraphBuilder methods to the textlayout module
