+++ render-wasm/src/shapes/text.rs
@@ -1,19 +1,14 @@
 use crate::{
     math::{Matrix, Rect},
-    render::{default_font, filters::compose_filters, DEFAULT_EMOJI_FONT},
-    textlayout::{auto_height, auto_width},
+    render::{default_font, DEFAULT_EMOJI_FONT},
+    textlayout::{auto_height, auto_width, paragraph_builders_from_text_content},
 };
-use skia_safe::{
-    self as skia,
-    paint::Paint,
-    textlayout::{ParagraphBuilder, ParagraphStyle},
-    ImageFilter, MaskFilter,
-};
+use skia_safe::{self as skia, paint::Paint, textlayout::ParagraphStyle, ImageFilter, MaskFilter};
 use std::collections::HashSet;
 
 use super::FontFamily;
-use crate::shapes::{self, merge_fills, set_paint_fill, Stroke, StrokeKind};
-use crate::utils::{get_fallback_fonts, get_font_collection, uuid_from_u32};
+use crate::shapes::{self, merge_fills};
+use crate::utils::uuid_from_u32;
 use crate::wasm::fills::parse_fills_from_bytes;
 use crate::Uuid;
 
@@ -61,6 +56,10 @@ impl TextContent {
         }
     }
 
+    pub fn bounds(&self) -> Rect {
+        self.bounds
+    }
+
     pub fn set_xywh(&mut self, x: f32, y: f32, w: f32, h: f32) {
         self.bounds = Rect::from_xywh(x, y, w, h);
     }
@@ -79,87 +78,9 @@ impl TextContent {
         self.paragraphs.push(paragraph);
     }
 
-    pub fn to_paragraph_builders(
-        &self,
-        blur: Option<&ImageFilter>,
-        blur_mask: Option<&MaskFilter>,
-    ) -> Vec<Vec<ParagraphBuilder>> {
-        let fonts = get_font_collection();
-        let fallback_fonts = get_fallback_fonts();
-        let mut paragraph_group = Vec::new();
-
-        for paragraph in &self.paragraphs {
-            let paragraph_style = paragraph.paragraph_to_style();
-            let mut builder = ParagraphBuilder::new(&paragraph_style, fonts);
-            for leaf in &paragraph.children {
-                let text_style = leaf.to_style(&self.bounds, fallback_fonts, blur, blur_mask);
-                let text = leaf.apply_text_transform();
-                builder.push_style(&text_style);
-                builder.add_text(&text);
-            }
-            paragraph_group.push(vec![builder]);
-        }
-
-        paragraph_group
-    }
-
-    pub fn to_stroke_paragraph_builders(
-        &self,
-        stroke: &Stroke,
-        bounds: &Rect,
-        blur: Option<&ImageFilter>,
-        blur_mask: Option<&MaskFilter>,
-        count_inner_strokes: usize,
-    ) -> Vec<Vec<ParagraphBuilder>> {
-        let fallback_fonts = get_fallback_fonts();
-        let fonts = get_font_collection();
-        let mut paragraph_group = Vec::new();
-
-        for paragraph in &self.paragraphs {
-            let mut stroke_paragraphs_map: std::collections::HashMap<usize, ParagraphBuilder> =
-                std::collections::HashMap::new();
-
-            for leaf in paragraph.children.iter() {
-                let mut text_paint = merge_fills(&leaf.fills, *bounds);
-                if let Some(blur_mask) = blur_mask {
-                    text_paint.set_mask_filter(blur_mask.clone());
-                }
-                let stroke_paints = get_text_stroke_paints(
-                    stroke,
-                    bounds,
-                    &text_paint,
-                    blur,
-                    blur_mask,
-                    count_inner_strokes,
-                );
-                let text: String = leaf.apply_text_transform();
-
-                for (paint_idx, stroke_paint) in stroke_paints.iter().enumerate() {
-                    let builder = stroke_paragraphs_map.entry(paint_idx).or_insert_with(|| {
-                        let paragraph_style = paragraph.paragraph_to_style();
-                        ParagraphBuilder::new(&paragraph_style, fonts)
-                    });
-                    let stroke_paint = stroke_paint.clone();
-                    let stroke_style =
-                        leaf.to_stroke_style(&stroke_paint, fallback_fonts, blur, blur_mask);
-                    builder.push_style(&stroke_style);
-                    builder.add_text(&text);
-                }
-            }
-
-            let stroke_paragraphs: Vec<ParagraphBuilder> = (0..stroke_paragraphs_map.len())
-                .map(|i| stroke_paragraphs_map.remove(&i).unwrap())
-                .collect();
-
-            paragraph_group.push(stroke_paragraphs);
-        }
-
-        paragraph_group
-    }
-
     pub fn width(&self) -> f32 {
         if self.grow_type() == GrowType::AutoWidth {
-            let mut paragraphs = self.to_paragraph_builders(None, None);
+            let mut paragraphs = paragraph_builders_from_text_content(self, None, None);
             auto_width(&mut paragraphs, f32::MAX).ceil()
         } else {
             self.bounds.width()
@@ -176,7 +97,7 @@ impl TextContent {
 
     pub fn visual_bounds(&self) -> (f32, f32) {
         let paragraph_width = self.width();
-        let mut paragraphs = self.to_paragraph_builders(None, None);
+        let mut paragraphs = paragraph_builders_from_text_content(self, None, None);
         let paragraph_height = auto_height(&mut paragraphs, paragraph_width);
         (paragraph_width, paragraph_height)
     }
@@ -267,7 +188,7 @@ impl Paragraph {
     }
 
     #[allow(dead_code)]
-    pub fn get_children(&self) -> &Vec<TextLeaf> {
+    pub fn children(&self) -> &Vec<TextLeaf> {
         &self.children
     }
 
@@ -434,6 +355,10 @@ impl TextLeaf {
             _ => skia::textlayout::TextDecoration::NO_DECORATION,
         });
         style
+    }
+
+    pub fn fills(&self) -> &Vec<shapes::Fill> {
+        &self.fills
     }
 
     fn serialized_font_family(&self) -> String {
@@ -675,99 +600,4 @@ impl From<&Vec<u8>> for RawTextData {
 
         Self { paragraph }
     }
-}
-
-fn get_text_stroke_paints(
-    stroke: &Stroke,
-    bounds: &Rect,
-    text_paint: &Paint,
-    blur: Option<&ImageFilter>,
-    blur_mask: Option<&MaskFilter>,
-    count_inner_strokes: usize,
-) -> Vec<Paint> {
-    let mut paints = Vec::new();
-
-    match stroke.kind {
-        StrokeKind::Inner => {
-            let shader = text_paint.shader();
-            let mut is_opaque = true;
-
-            if shader.is_some() {
-                is_opaque = shader.unwrap().is_opaque();
-            }
-
-            if is_opaque && count_inner_strokes == 1 {
-                let mut paint = text_paint.clone();
-                paint.set_style(skia::PaintStyle::Fill);
-                paint.set_anti_alias(true);
-                if let Some(blur) = blur {
-                    paint.set_image_filter(blur.clone());
-                }
-                paints.push(paint);
-                let mut paint = skia::Paint::default();
-                paint.set_style(skia::PaintStyle::Stroke);
-                paint.set_blend_mode(skia::BlendMode::SrcIn);
-                paint.set_anti_alias(true);
-                paint.set_stroke_width(stroke.width * 2.0);
-                set_paint_fill(&mut paint, &stroke.fill, bounds);
-                if let Some(blur) = blur {
-                    paint.set_image_filter(blur.clone());
-                }
-                paints.push(paint);
-            } else {
-                let mut paint = text_paint.clone();
-                paint.set_style(skia::PaintStyle::Fill);
-                paint.set_anti_alias(false);
-                set_paint_fill(&mut paint, &stroke.fill, bounds);
-                paints.push(paint);
-
-                let mut paint = skia::Paint::default();
-                let image_filter =
-                    skia_safe::image_filters::erode((stroke.width, stroke.width), None, None);
-
-                let filter = compose_filters(blur, image_filter.as_ref());
-                paint.set_image_filter(filter);
-                paint.set_anti_alias(false);
-                paint.set_blend_mode(skia::BlendMode::DstOut);
-                paints.push(paint);
-            }
-        }
-        StrokeKind::Center => {
-            let mut paint = skia::Paint::default();
-            paint.set_style(skia::PaintStyle::Stroke);
-            paint.set_anti_alias(true);
-            paint.set_stroke_width(stroke.width);
-
-            set_paint_fill(&mut paint, &stroke.fill, bounds);
-            if let Some(blur) = blur {
-                paint.set_image_filter(blur.clone());
-            }
-
-            paints.push(paint);
-        }
-        StrokeKind::Outer => {
-            let mut paint = skia::Paint::default();
-            paint.set_style(skia::PaintStyle::Stroke);
-            paint.set_blend_mode(skia::BlendMode::DstOver);
-            paint.set_anti_alias(true);
-            paint.set_stroke_width(stroke.width * 2.0);
-            set_paint_fill(&mut paint, &stroke.fill, bounds);
-            if let Some(blur_mask) = blur_mask {
-                paint.set_mask_filter(blur_mask.clone());
-            }
-            paints.push(paint);
-
-            let mut paint = skia::Paint::default();
-            paint.set_style(skia::PaintStyle::Fill);
-            paint.set_blend_mode(skia::BlendMode::Clear);
-            paint.set_color(skia::Color::TRANSPARENT);
-            paint.set_anti_alias(true);
-            if let Some(blur_mask) = blur_mask {
-                paint.set_mask_filter(blur_mask.clone());
-            }
-            paints.push(paint);
-        }
-    }
-
-    paints
 }
